+++ inbas-reporting-tool/src/main/java/cz/cvut/kbss/reporting/persistence/dao/OccurrenceReportDao.java
@@ -7,7 +7,9 @@ import cz.cvut.kbss.reporting.model.Occurrence;
 import cz.cvut.kbss.reporting.model.OccurrenceReport;
 import cz.cvut.kbss.reporting.model.Vocabulary;
 import cz.cvut.kbss.reporting.persistence.util.OrphanRemover;
+import cz.cvut.kbss.reporting.service.event.InvalidateCacheEvent;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.ApplicationListener;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
@@ -22,22 +24,27 @@ import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Repository
-public class OccurrenceReportDao extends BaseReportDao<OccurrenceReport> {
+public class OccurrenceReportDao extends BaseReportDao<OccurrenceReport>
+        implements ApplicationListener<InvalidateCacheEvent> {
 
     private static final String SELECT = "SELECT DISTINCT ?x WHERE { ";
     private static final String WHERE_CONDITION = "?x a ?type ; " +
             "?hasKey ?key ;" +
-            "?hasRevision ?revision ;" +
             "?hasAuthor ?author ;" +
+            "?hasCreatedDate ?created ;" +
             "?hasOccurrence ?occurrence ." +
             "OPTIONAL { ?x ?hasSeverity ?severity . }" +
             "OPTIONAL { ?x ?hasLastEditor ?lastEditor . }" +
-            "?occurrence ?hasStartTime ?startTime ;" +
-            "?hasEventType ?occurrenceCategory ." +
+            "OPTIONAL { ?x ?hasLastEditedDate ?lastModified . }" +
+            "?occurrence ?hasEventType ?occurrenceCategory ." +
             "FILTER NOT EXISTS { " +
             "?y a ?type . " +
-            "?x ?hasNext ?y . }";
-    private static final String QUERY_TAIL = "} ORDER BY DESC(?startTime) DESC(?revision) LIMIT ?limit OFFSET ?offset";
+            "?x ?hasNext ?y . }" +
+            "BIND (IF (BOUND(?lastModified), ?lastModified, ?created) AS ?edited)";
+    private static final String QUERY_TAIL = "} ORDER BY DESC(?edited) LIMIT ?limit OFFSET ?offset";
+
+    // Current count of latest revisions of reports
+    private volatile long reportCount = -1;
 
     private final OccurrenceDao occurrenceDao;
 
@@ -71,20 +78,20 @@ public class OccurrenceReportDao extends BaseReportDao<OccurrenceReport> {
                         cz.cvut.kbss.reporting.model.reportlist.OccurrenceReport.class)
                 .setParameter("type", typeUri)
                 .setParameter("hasKey", URI.create(Vocabulary.s_p_has_key))
-                .setParameter("hasRevision", URI.create(Vocabulary.s_p_has_revision))
                 .setParameter("hasSeverity", URI.create(Vocabulary.s_p_has_severity_assessment))
                 .setParameter("hasLastEditor", URI.create(Vocabulary.s_p_has_last_editor))
                 .setParameter("hasAuthor", URI.create(Vocabulary.s_p_has_author))
                 .setParameter("hasOccurrence", URI.create(Vocabulary.s_p_documents))
-                .setParameter("hasStartTime", URI.create(Vocabulary.s_p_has_start_time))
                 .setParameter("hasEventType", URI.create(Vocabulary.s_p_has_event_type))
                 .setParameter("hasNext", URI.create(Vocabulary.s_p_has_next_revision))
+                .setParameter("hasCreatedDate", URI.create(Vocabulary.s_p_created))
+                .setParameter("hasLastEditedDate", URI.create(Vocabulary.s_p_modified))
                 .setUntypedParameter("limit", pageSpec.getPageSize())
                 .setUntypedParameter("offset", pageSpec.getPageSize() * pageSpec.getPageNumber())
                 .getResultList();
         return new PageImpl<>(
                 res.stream().map(cz.cvut.kbss.reporting.model.reportlist.OccurrenceReport::toOccurrenceReport)
-                   .collect(Collectors.toList()), pageSpec, 0L);
+                   .collect(Collectors.toList()), pageSpec, getReportCount(em));
     }
 
     private String buildQuery(Collection<ReportFilter> filters) {
@@ -100,6 +107,31 @@ public class OccurrenceReportDao extends BaseReportDao<OccurrenceReport> {
         return sb.toString();
     }
 
+    /**
+     * Gets total count of latest revisions of all reports in the repository.
+     *
+     * @param em EntityManager
+     * @return Report count
+     */
+    private long getReportCount(EntityManager em) {
+        if (reportCount == -1) {
+            this.reportCount = em.createNativeQuery("SELECT (count(?x) as ?cnt) WHERE {" +
+                    "?x a ?type ;" +
+                    "FILTER NOT EXISTS {\n" +
+                    "?y a ?type ." +
+                    "?x ?hasNext ?y . }" +
+                    "}", Integer.class) // Count returns an xsd:int (as per SPARQL specification)
+                                 .setParameter("type", typeUri)
+                                 .setParameter("hasNext", URI.create(Vocabulary.s_p_has_next_revision))
+                                 .getSingleResult();
+        }
+        return reportCount;
+    }
+
+    private void resetReportCount() {
+        this.reportCount = -1;
+    }
+
     @Override
     protected void persist(OccurrenceReport entity, EntityManager em) {
         assert entity != null;
@@ -107,6 +139,7 @@ public class OccurrenceReportDao extends BaseReportDao<OccurrenceReport> {
             occurrenceDao.persist(entity.getOccurrence(), em);
         }
         super.persist(entity, em);
+        resetReportCount();
     }
 
     @Override
@@ -127,6 +160,7 @@ public class OccurrenceReportDao extends BaseReportDao<OccurrenceReport> {
     protected void remove(OccurrenceReport entity, EntityManager em) {
         occurrenceDao.remove(entity.getOccurrence(), em);
         super.remove(entity, em);
+        resetReportCount();
     }
 
     /**
@@ -145,7 +179,7 @@ public class OccurrenceReportDao extends BaseReportDao<OccurrenceReport> {
                             "?x a ?type ;" +
                             "?documents ?occurrence . }",
                     OccurrenceReport.class)
-                     .setParameter("type", typeIri)
+                     .setParameter("type", typeUri)
                      .setParameter("documents", URI.create(Vocabulary.s_p_documents))
                      .setParameter("occurrence", occurrence.getUri())
                      .getSingleResult();
@@ -154,5 +188,10 @@ public class OccurrenceReportDao extends BaseReportDao<OccurrenceReport> {
         } finally {
             em.close();
         }
+    }
+
+    @Override
+    public void onApplicationEvent(InvalidateCacheEvent invalidateCacheEvent) {
+        resetReportCount();
     }
 }