+++ inbas-reporting-tool/src/main/java/cz/cvut/kbss/reporting/service/MainReportService.java
@@ -6,8 +6,8 @@ import cz.cvut.kbss.reporting.exception.NotFoundException;
 import cz.cvut.kbss.reporting.exception.UnsupportedReportTypeException;
 import cz.cvut.kbss.reporting.filter.ReportFilter;
 import cz.cvut.kbss.reporting.model.*;
-import cz.cvut.kbss.reporting.model.util.DocumentDateAndRevisionComparator;
 import cz.cvut.kbss.reporting.model.util.EntityToOwlClassMapper;
+import cz.cvut.kbss.reporting.model.util.ReportLastModifiedComparator;
 import cz.cvut.kbss.reporting.persistence.dao.ReportDao;
 import cz.cvut.kbss.reporting.service.data.AttachmentService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -60,7 +60,7 @@ public class MainReportService implements ReportBusinessService {
         final List<LogicalDocument> reports = new ArrayList<>();
         services.values().forEach(service -> reports.addAll(service.findAll()));
         return reports.stream().map(LogicalDocument::toReportDto)
-                      .sorted(new DocumentDateAndRevisionComparator())
+                      .sorted(new ReportLastModifiedComparator())
                       .collect(Collectors.toList());
     }
 
@@ -68,12 +68,18 @@ public class MainReportService implements ReportBusinessService {
     public Page<ReportDto> findAll(Pageable pageSpec, Collection<ReportFilter> filters) {
         final List<LogicalDocument> reports = new ArrayList<>();
         // Combine reports from all services
-        services.values().forEach(service -> reports.addAll(service.findAll(pageSpec, filters).getContent()));
+        long totalOfTotals = 0;
+        for (BaseReportService<? extends LogicalDocument> service : services.values()) {
+            final Page<? extends LogicalDocument> page = service.findAll(pageSpec, filters);
+            reports.addAll(page.getContent());
+            totalOfTotals += page.getTotalElements();
+        }
         final List<ReportDto> result = reports.stream().map(LogicalDocument::toReportDto)
-                                              .sorted(new DocumentDateAndRevisionComparator())
+                                              .sorted(new ReportLastModifiedComparator())
                                               .collect(Collectors.toList());
         // And return corresponding page
-        return new PageImpl<>(result.subList(0, Math.min(result.size(), pageSpec.getPageSize())), pageSpec, 0L);
+        return new PageImpl<>(result.subList(0, Math.min(result.size(), pageSpec.getPageSize())), pageSpec,
+                totalOfTotals);
     }
 
     @Override