+++ inbas-reporting-tool/src/main/webapp/tests/__tests__/FiltersTest.js
@@ -1,5 +1,3 @@
-'use strict';
-
 import React from "react";
 import TestUtils from "react-addons-test-utils";
 import Actions from "../../js/actions/Actions";
@@ -13,7 +11,7 @@ import Filters from "../../js/components/filter/Filters";
 
 describe('Filters', () => {
 
-    var onChange, onResetFilters,
+    let onChange, onResetFilters,
         reports, count;
 
     beforeEach(() => {
@@ -25,27 +23,27 @@ describe('Filters', () => {
     });
 
     it('adds default filter option - any - to the rendered select', () => {
-        var categories = getCategoriesAsJsonLd();
+        const categories = getCategoriesAsJsonLd();
         spyOn(OptionsStore, 'getOptions').and.returnValue(categories);
-        var component = Environment.render(<Filters filters={{}} data={reports} onChange={onChange}
+        const component = Environment.render(<Filters filters={{}} data={reports} onChange={onChange}
                                                     onResetFilters={onResetFilters}/>),
             options = component.state[Constants.FILTERS[0].options];
 
-        var defaultOption = options[0];
+        const defaultOption = options[0];
         expect(defaultOption.value).toEqual(Constants.FILTER_DEFAULT);
         expect(defaultOption.label).toEqual(messages['reports.filter.type.all']);
     });
 
     it('generates options for occurrence categories from categories existing in the data', () => {
-        var categories = getCategoriesAsJsonLd();
+        const categories = getCategoriesAsJsonLd();
         spyOn(OptionsStore, 'getOptions').and.returnValue(categories);
-        var component = Environment.render(<Filters filters={{}} data={reports} onChange={onChange}
+        const component = Environment.render(<Filters filters={{}} data={reports} onChange={onChange}
                                                     onResetFilters={onResetFilters}/>),
             options = component.state[Constants.FILTERS[0].options],
             categoriesUsed = getCategoriesUsedByReports(categories);
 
         expect(options.length).toEqual(categoriesUsed.length + 1);
-        for (var i = 0, len = categoriesUsed.length; i < len; i++) {
+        for (let i = 0, len = categoriesUsed.length; i < len; i++) {
             expect(options[i + 1].value).toEqual(categoriesUsed[i]['@id']);
             expect(options[i + 1].label).toEqual(categoriesUsed[i][Vocabulary.RDFS_LABEL]);
             expect(options[i + 1].title).toEqual(categoriesUsed[i][Vocabulary.RDFS_COMMENT]);
@@ -53,9 +51,9 @@ describe('Filters', () => {
     });
 
     function getCategoriesAsJsonLd() {
-        var categories = Generator.getCategories();
+        const categories = Generator.getCategories();
         return categories.map(item => {
-            var res = {};
+            const res = {};
             res['@id'] = item.id;
             res[Vocabulary.RDFS_LABEL] = item.name;
             res[Vocabulary.RDFS_COMMENT] = item.description;
@@ -64,11 +62,11 @@ describe('Filters', () => {
     }
 
     function getCategoriesUsedByReports(categories) {
-        var cat,
+        let cat,
             categoriesUsed = [];
-        for (var i = 0, len = categories.length; i < len; i++) {
+        for (let i = 0, len = categories.length; i < len; i++) {
             cat = categories[i];
-            for (var j = 0; j < count; j++) {
+            for (let j = 0; j < count; j++) {
                 if (categoriesUsed.indexOf(cat) === -1 &&
                     (cat['@id'] === reports[j].occurrenceCategory || Array.isArray(reports[j].occurrenceCategory) &&
                     reports[j].occurrenceCategory.indexOf(cat['@id']) !== -1)) {
@@ -81,25 +79,25 @@ describe('Filters', () => {
     }
 
     it('generates options for occurrence categories existing in data when multiple categories per record exist', () => {
-        var multipleCats = false;
-        for (var i = 0, len = reports.length; i < len; i++) {
+        let multipleCats = false;
+        for (let i = 0, len = reports.length; i < len; i++) {
             if (!multipleCats || Generator.getRandomBoolean()) {
                 reports[i].occurrenceCategory = [reports[i].occurrenceCategory, Generator.randomCategory().id];
                 multipleCats = true;
             }
         }
-        var categories = getCategoriesAsJsonLd();
+        const categories = getCategoriesAsJsonLd();
         spyOn(OptionsStore, 'getOptions').and.returnValue(categories);
-        var component = Environment.render(<Filters filters={{}} data={reports} onChange={onChange}
+        const component = Environment.render(<Filters filters={{}} data={reports} onChange={onChange}
                                                     onResetFilters={onResetFilters}/>),
             options = component.state[Constants.FILTERS[0].options],
             categoriesUsed = getCategoriesUsedByReports(categories);
 
         expect(options.length).toEqual(categoriesUsed.length + 1);
-        var found;
-        for (i = 0, len = categoriesUsed.length; i < len; i++) {
+        let found;
+        for (let i = 0, len = categoriesUsed.length; i < len; i++) {
             found = false;
-            for (var j = 1, lenn = options.length; j < lenn; j++) {
+            for (let j = 1, lenn = options.length; j < lenn; j++) {
                 if (categoriesUsed[i]['@id'] === options[j].value) {
                     expect(options[j].label).toEqual(categoriesUsed[i][Vocabulary.RDFS_LABEL]);
                     expect(options[j].title).toEqual(categoriesUsed[i][Vocabulary.RDFS_COMMENT]);
@@ -112,16 +110,16 @@ describe('Filters', () => {
     });
 
     it('renders select with options corresponding to all unique categories used by data', () => {
-        var categories = getCategoriesAsJsonLd();
+        const categories = getCategoriesAsJsonLd();
         spyOn(OptionsStore, 'getOptions').and.returnValue(categories);
-        var component = Environment.render(<Filters filters={{}} data={reports} onChange={onChange}
+        const component = Environment.render(<Filters filters={{}} data={reports} onChange={onChange}
                                                     onResetFilters={onResetFilters}/>),
             categoriesUsed = getCategoriesUsedByReports(categories);
 
-        var select = TestUtils.scryRenderedComponentsWithType(component, require('../../js/components/Select'))[0],
+        const select = TestUtils.scryRenderedComponentsWithType(component, require('../../js/components/Select'))[0],
             options = TestUtils.scryRenderedDOMComponentsWithTag(select, 'option');
         expect(options.length).toEqual(categoriesUsed.length + 1);
-        for (var i = 0, len = categoriesUsed.length; i < len; i++) {
+        for (let i = 0, len = categoriesUsed.length; i < len; i++) {
             expect(options[i + 1].value).toEqual(categoriesUsed[i]['@id']);
             expect(options[i + 1].textContent).toEqual(categoriesUsed[i][Vocabulary.RDFS_LABEL]);
             expect(options[i + 1].title).toEqual(categoriesUsed[i][Vocabulary.RDFS_COMMENT]);
@@ -129,10 +127,10 @@ describe('Filters', () => {
     });
 
     it('sets value of a filter to the value specified in props', () => {
-        var categories = getCategoriesAsJsonLd(),
+        const categories = getCategoriesAsJsonLd(),
             value = reports[Generator.getRandomInt(reports.length)].occurrenceCategory;
         spyOn(OptionsStore, 'getOptions').and.returnValue(categories);
-        var component = Environment.render(<Filters filters={{occurrenceCategory: value}} data={reports}
+        const component = Environment.render(<Filters filters={{occurrenceCategory: value}} data={reports}
                                                     onChange={onChange} onResetFilters={onResetFilters}/>),
 
             select = TestUtils.scryRenderedDOMComponentsWithTag(component, 'select')[0];
@@ -140,15 +138,26 @@ describe('Filters', () => {
     });
 
     it('notifies owner about change in the filter value', () => {
-        var categories = getCategoriesAsJsonLd(),
+        const categories = getCategoriesAsJsonLd(),
             value = reports[Generator.getRandomInt(reports.length)].occurrenceCategory;
         spyOn(OptionsStore, 'getOptions').and.returnValue(categories);
-        var component = Environment.render(<Filters filters={{}} data={reports}
+        const component = Environment.render(<Filters filters={{}} data={reports}
                                                     onChange={onChange} onResetFilters={onResetFilters}/>),
 
             select = TestUtils.scryRenderedDOMComponentsWithTag(component, 'select')[0];
         select.value = value;
         TestUtils.Simulate.change(select);
         expect(onChange).toHaveBeenCalledWith({occurrenceCategory: value});
+    });
+
+    it('does not render filter if there is only one value', () => {
+        const categories = getCategoriesAsJsonLd();
+        reports.forEach(r => r.phase = 'http://onto.fel.cvut.cz/ontologies/documentation/processed');
+        spyOn(OptionsStore, 'getOptions').and.returnValue(categories);
+        const component = Environment.render(<Filters filters={{}} data={reports}
+                                                      onChange={onChange} onResetFilters={onResetFilters}/>),
+            selects = TestUtils.scryRenderedDOMComponentsWithTag(component, 'select');
+        expect(selects.length).toEqual(1);
+        expect(selects[0].name).toEqual(Constants.FILTERS[0].path);
     });
 });