+++ inbas-reporting-tool/src/test/java/cz/cvut/kbss/reporting/persistence/dao/OccurrenceReportDaoTest.java
@@ -7,7 +7,9 @@ import cz.cvut.kbss.reporting.environment.generator.OccurrenceReportGenerator;
 import cz.cvut.kbss.reporting.environment.util.Environment;
 import cz.cvut.kbss.reporting.filter.*;
 import cz.cvut.kbss.reporting.model.*;
+import cz.cvut.kbss.reporting.model.util.ReportLastModifiedComparator;
 import cz.cvut.kbss.reporting.persistence.BaseDaoTestRunner;
+import cz.cvut.kbss.reporting.service.event.InvalidateCacheEvent;
 import org.hamcrest.Matchers;
 import org.hamcrest.number.OrderingComparison;
 import org.junit.Before;
@@ -17,9 +19,11 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 
+import java.lang.reflect.Field;
 import java.net.URI;
 import java.util.*;
 import java.util.function.Consumer;
+import java.util.function.Function;
 import java.util.stream.Collectors;
 
 import static cz.cvut.kbss.reporting.environment.util.TestUtils.verifyAllInstancesRemoved;
@@ -66,7 +70,7 @@ public class OccurrenceReportDaoTest extends BaseDaoTestRunner {
     private OccurrenceReport report(boolean withFactorGraph) {
         final OccurrenceReport report =
                 withFactorGraph ? OccurrenceReportGenerator.generateOccurrenceReportWithFactorGraph() :
-                        OccurrenceReportGenerator.generateOccurrenceReport(true);
+                OccurrenceReportGenerator.generateOccurrenceReport(true);
         report.setAuthor(author);
         return report;
     }
@@ -346,6 +350,7 @@ public class OccurrenceReportDaoTest extends BaseDaoTestRunner {
     public void pageableFindAllReturnsReportsOnMatchingPage() {
         final List<OccurrenceReport> reports = generateReports();
         occurrenceReportDao.persist(reports);
+        reports.sort(Comparator.comparing(OccurrenceReport::getDateCreated).reversed());
         final int pageSize = 5;
         final int pageNo = 1;   // Page numbers start from 0
         final Pageable pageReq = PageRequest.of(pageNo, pageSize);
@@ -361,7 +366,10 @@ public class OccurrenceReportDaoTest extends BaseDaoTestRunner {
     private List<OccurrenceReport> generateReports() {
         final int reportCount = 15;
         final List<OccurrenceReport> reports = OccurrenceReportGenerator.generateReports(true, reportCount);
-        reports.forEach(r -> r.setAuthor(author));
+        for (int i = 0; i < reports.size(); i++) {
+            reports.get(i).setAuthor(author);
+            reports.get(i).setDateCreated(new Date(System.currentTimeMillis() - i * 1000));
+        }
         return reports;
     }
 
@@ -400,8 +408,8 @@ public class OccurrenceReportDaoTest extends BaseDaoTestRunner {
         final List<OccurrenceReport> matching = allReports.stream().filter(r -> Generator.randomBoolean())
                                                           .peek(transformation)
                                                           .sorted(Comparator.comparing(
-                                                                  (OccurrenceReport r) -> r.getOccurrence()
-                                                                                           .getStartTime()).reversed())
+                                                                  (Function<OccurrenceReport, Date>) AbstractReport::getDateCreated)
+                                                                            .reversed())
                                                           .collect(Collectors.toList());
         occurrenceReportDao.persist(allReports);
         return matching;
@@ -584,5 +592,93 @@ public class OccurrenceReportDaoTest extends BaseDaoTestRunner {
                 occurrenceReportDao.findAll(PageRequest.of(0, 1), Collections.emptyList());
         final OccurrenceReport result = updated.getContent().get(0);
         assertEquals(newSummary, result.getSummary());
+    }
+
+    @Test
+    public void findAllReturnsPageWithCorrectTotalReportCount() {
+        final List<OccurrenceReport> reports = generateReports();
+        occurrenceReportDao.persist(reports);
+        final Pageable req = PageRequest.of(0, reports.size() - 1);
+        final Page<OccurrenceReport> result = occurrenceReportDao.findAll(req, Collections.emptyList());
+        assertEquals(reports.size(), result.getTotalElements());
+    }
+
+    @Test
+    public void findAllReturnsCorrectTotalCountAfterReportPersist() {
+        final List<OccurrenceReport> reports = generateReports();
+        occurrenceReportDao.persist(reports);
+        final Pageable req = PageRequest.of(0, reports.size() - 1);
+        final Page<OccurrenceReport> intermediate = occurrenceReportDao.findAll(req, Collections.emptyList());
+        assertEquals(reports.size(), intermediate.getTotalElements());
+        final OccurrenceReport added = report(false);
+        occurrenceReportDao.persist(added);
+        final Page<OccurrenceReport> result = occurrenceReportDao.findAll(req, Collections.emptyList());
+        assertEquals(intermediate.getTotalElements() + 1, result.getTotalElements());
+    }
+
+    @Test
+    public void findAllReturnsCorrectTotalCountAfterReportRemoval() {
+        final List<OccurrenceReport> reports = generateReports();
+        occurrenceReportDao.persist(reports);
+        final Pageable req = PageRequest.of(0, reports.size() / 2);
+        final Page<OccurrenceReport> intermediate = occurrenceReportDao.findAll(req, Collections.emptyList());
+        assertEquals(reports.size(), intermediate.getTotalElements());
+        occurrenceReportDao.remove(reports.get(Generator.randomIndex(reports)));
+        final Page<OccurrenceReport> result = occurrenceReportDao.findAll(req, Collections.emptyList());
+        assertEquals(intermediate.getTotalElements() - 1, result.getTotalElements());
+    }
+
+    @Test
+    public void invalidateCacheCausesReportCountReset() throws Exception {
+        final List<OccurrenceReport> reports = generateReports();
+        occurrenceReportDao.persist(reports);
+        occurrenceReportDao.findAll(PageRequest.of(0, Integer.MAX_VALUE), Collections.emptyList());
+        final Field countField = OccurrenceReportDao.class.getDeclaredField("reportCount");
+        countField.setAccessible(true);
+        final long count = (long) countField.get(occurrenceReportDao);
+        assertEquals(reports.size(), count);
+        occurrenceReportDao.onApplicationEvent(new InvalidateCacheEvent(this));
+        assertEquals(-1, (long) countField.get(occurrenceReportDao));
+    }
+
+    @Test
+    public void findAllReturnsReportsOrderedByLastModifiedDateDescending() {
+        final List<OccurrenceReport> reports = generateReports();
+        Collections.shuffle(reports);
+        for (int i = 0; i < reports.size(); i++) {
+            reports.get(i).setDateCreated(new Date(System.currentTimeMillis() - 24 * 3600 * 1000));
+            reports.get(i).setLastModified(new Date(System.currentTimeMillis() - i * 10000));
+        }
+        Collections.shuffle(reports);
+        occurrenceReportDao.persist(reports);
+        reports.sort(Comparator.comparing(OccurrenceReport::getLastModified).reversed());
+
+        final Page<OccurrenceReport> result = occurrenceReportDao
+                .findAll(PageRequest.of(0, Integer.MAX_VALUE), Collections.emptyList());
+        assertEquals(reports.size(), result.getNumberOfElements());
+        for (int i = 0; i < reports.size(); i++) {
+            assertEquals(reports.get(i).getUri(), result.getContent().get(i).getUri());
+        }
+    }
+
+    @Test
+    public void findAllReturnsReportsOrderedByLastModifiedOrDateCreatedDescending() {
+        final List<OccurrenceReport> reports = generateReports();
+        Collections.shuffle(reports);
+        for (int i = 0; i < reports.size(); i++) {
+            if (Generator.randomBoolean()) {
+                reports.get(i).setLastModified(new Date(System.currentTimeMillis() + (i + 1) * 10000));
+            }
+        }
+        Collections.shuffle(reports);
+        occurrenceReportDao.persist(reports);
+        reports.sort(new ReportLastModifiedComparator());
+
+        final Page<OccurrenceReport> result = occurrenceReportDao
+                .findAll(PageRequest.of(0, Integer.MAX_VALUE), Collections.emptyList());
+        assertEquals(reports.size(), result.getNumberOfElements());
+        for (int i = 0; i < reports.size(); i++) {
+            assertEquals(reports.get(i).getUri(), result.getContent().get(i).getUri());
+        }
     }
 }